Address CVE-2018-10860
perl-archive-zip is vulnerable to a directory traversal in Archive::Zip. It
was found that the Archive::Zip module did not properly sanitize paths
while extracting zip files. An attacker able to provide a specially crafted
archive for processing could use this flaw to write or overwrite arbitrary
files in the context of the perl interpreter.

Signed-off-by: Peter O'Connor <peter@solus-project.com>